# feat: Add observability tooling and CI/CD infrastructure