--- a/.github/ISSUE_TEMPLATE/bug.md
+++ b/.github/ISSUE_TEMPLATE/bug.md
@@ -1,0 +1,41 @@
+---
+name: Bug Report
+about: Report a bug in the network monitor
+title: '[BUG] '
+labels: bug
+assignees: ''
+---
+
+## 🐛 Bug Description
+
+A clear description of what the bug is.
+
+## 🔄 Steps to Reproduce
+
+1. Go to '...'
+2. Click on '...'
+3. Scroll down to '...'
+4. See error
+
+## ✅ Expected Behavior
+
+What should happen.
+
+## ❌ Actual Behavior
+
+What actually happens.
+
+## 🖼️ Screenshots
+
+If applicable, add screenshots.
+
+## 🔧 Environment
+
+- Browser: [e.g., Chrome, Firefox]
+- OS: [e.g., macOS, Windows]
+- Version: [e.g., 1.0.0]
+
+## 📋 Additional Context
+
+Any other context about the problem.
+