--- a/.github/ISSUE_TEMPLATE/feature.md
+++ b/.github/ISSUE_TEMPLATE/feature.md
@@ -1,0 +1,32 @@
+---
+name: Feature Request
+about: Suggest a new feature for the network monitor
+title: '[FEATURE] '
+labels: enhancement
+assignees: ''
+---
+
+## 📋 Feature Description
+
+A clear description of the feature you want to implement.
+
+## 🎯 Motivation
+
+Why is this feature needed? What problem does it solve?
+
+## 💡 Proposed Solution
+
+Describe how you want to implement this feature.
+
+## 📝 Implementation Checklist
+
+- [ ] Design/Architecture
+- [ ] Implementation
+- [ ] Tests
+- [ ] Documentation
+- [ ] Code review
+
+## 🔗 Related Issues
+
+List any related issues or dependencies.
+