--- a/.github/ISSUE_TEMPLATE/task.md
+++ b/.github/ISSUE_TEMPLATE/task.md
@@ -1,0 +1,28 @@
+---
+name: Task
+about: General development task
+title: '[TASK] '
+labels: task
+assignees: ''
+---
+
+## 📋 Task Description
+
+What needs to be done?
+
+## 🎯 Acceptance Criteria
+
+- [ ] Criterion 1
+- [ ] Criterion 2
+- [ ] Criterion 3
+
+## 📝 Notes
+
+Any additional information or context.
+
+## 🔗 Related
+
+- Related issues: #
+- Depends on: #
+- Blocks: #
+