… .github/workflows/README.md
… .github/workflows/pr-automation.yml
… .github/workflows/status-checks.yml
… AGENTS.md
… GEMINI.md
… TASK-MANAGEMENT.md
… apps/api/API-QUICK-START.md
… apps/api/COLLECTIONS-GUIDE.md
… apps/api/COMPLETE-SERVER-STACK.md
… apps/api/COMPLETE.md
… apps/api/DOCUMENTATION-SUMMARY.md
… apps/api/DRY-ARCHITECTURE.md
… apps/api/FRAMEWORK-COMPARISON.md
… apps/api/HONO-MIGRATION.md
… apps/api/IMPLEMENTATION-SUMMARY.md
… apps/api/START-HERE.md
… apps/api/VISUAL-GUIDE.md
… apps/api/WHY-DIFFERENT-STACKS.md
… apps/web/TRPC-API-REFERENCE.md
… apps/web/TRPC-ARCHITECTURE.md
… apps/web/package.json
… bun.lock
… docs/PLG-STACK-IMPLEMENTATION.md
… docs/README.MD