- Add Storybook for SolidJS with testing integration (vitest, playwright)
- Add solid-primitives packages (storage, event-listener, timer, media, scheduled)
- Set up GitHub Issues workflow with MCP integration for Cursor AI
- Configure GitHub Actions CI/CD with automatic PR reviews
- Set up Gemini Code Assist for automated code reviews
- Create PLG Stack implementation plan (Prometheus, Loki, Grafana)
- Add comprehensive task management documentation
- Create issue templates (bug, feature, task)
- Add CI check scripts and feature branch workflow
- Clean up outdated documentation files

Created GitHub Issues:
- #2-6: Storybook implementation tasks
- #7-15: PLG Stack observability epic and implementation tasks

This establishes the foundation for:
- Component documentation and testing with Storybook
- Full observability stack with metrics, logs, and dashboards
- Automated CI/CD workflow with AI-assisted code reviews
- Task management integrated with Cursor AI via MCP