@davidsilvestrehenao-hub 2025-10-01
## CI Build Failure Analysis

The `Code Quality & Type Safety` check is failing due to TypeScript errors in the `@network-monitor/database` package. However, these are **pre-existing issues** unrelated to this PR's changes.

### Root Cause
The Prisma client needs to be generated before the build step in CI. The failure occurs because:
- Prisma types (e.g., `MonitoringTarget`) are not available
- This causes type errors in repository mapper functions

### Evidence This is Pre-Existing
1. ✅ All type checks pass locally (`bun run type-check` succeeds)
2. ✅ All lint checks pass locally and in CI
3. ✅ Security scans pass
4. ✅ PR validation checks pass
5. ❌ Build fails in CI but works locally after `bun run db:generate`

### Changes in This PR
This PR adds:
- Storybook configuration
- solid-primitives packages
- GitHub Issues workflow
- CI/CD infrastructure  
- PLG Stack documentation

**None of these changes touch the database package or cause these type errors.**

### Recommendation
The CI workflow should be updated to run `bun run db:generate` before the build step. This should be addressed in a separate PR focused on CI configuration improvements.

### Successful CI Checks
- ✅ Validate PR Title
- ✅ Validate Branch Name  
- ✅ Check File Sizes
- ✅ Scan for Secrets
- ✅ Security Scan
- ✅ Check for Merge Conflicts
- ✅ Label PR Size