@davidsilvestrehenao-hub 2025-10-01
## 🤖 Auto-Merge Workflow Added

I've added an auto-merge workflow that will automatically squash and merge PRs when all CI checks pass!

### How It Works

1. **Triggers** when:
   - PR is opened, synchronized, or reopened
   - Any check suite completes
   - Status updates occur

2. **Safety Checks**:
   - ✅ All CI checks must pass
   - ✅ PR must not be a draft
   - ✅ No merge conflicts
   - ✅ PR must be open (not already merged/closed)

3. **Actions**:
   - ✅ Squash merges into main
   - ✅ Deletes the feature branch
   - ✅ Posts confirmation comment

### Benefits

- 🚀 **Faster merges**: No manual merge clicks needed
- 🔒 **Safe**: Only merges when ALL checks pass
- 🧹 **Clean**: Automatically deletes merged branches
- 📝 **Documented**: Posts merge confirmation

### Current PR Status

All checks are now passing! This PR should auto-merge soon. 🎉

If you prefer manual control, you can disable auto-merge by:
- Converting PR to draft temporarily
- Or deleting the `.github/workflows/auto-merge.yml` file