@davidsilvestrehenao-hub 2025-10-01
## ⚡ CI Performance Optimization

I've added comprehensive caching to speed up CI builds!

### Optimizations Added

1. **Bun Dependency Caching**
   - Caches `node_modules` and Bun's install cache
   - Key: Based on `bun.lockb` hash
   - Benefit: Skip dependency installation when lockfile unchanged

2. **Turbo Build Caching**
   - Caches `.turbo` directory
   - Benefit: Skip building unchanged packages (incremental builds)

3. **Prisma Client Caching**
   - Caches `node_modules/.prisma`
   - Key: Based on Prisma schema hash
   - Benefit: Skip Prisma generation when schema unchanged

4. **Shallow Git Fetch**
   - Uses `fetch-depth: 2` instead of full history
   - Benefit: Faster checkout

### Expected Performance Gains

| Step | Before | After (cached) | Speedup |
|------|--------|---------------|---------|
| Install dependencies | ~60s | ~5-10s | **6-12x faster** |
| Build packages | ~90s | ~20-30s | **3-4x faster** |
| **Total CI time** | **~3-4 min** | **~45-60s** | **4-5x faster** |

### First Run vs Subsequent Runs

- **First run**: Same speed as before (builds cache)
- **Subsequent runs**: 4-5x faster (uses cache)
- **After changing lockfile**: Rebuilds dependency cache
- **After changing schema**: Regenerates Prisma client
- **After changing code**: Only rebuilds affected packages (Turbo)

The next CI run will see the full benefits! 🚀